requirements.txt
einops
timm==1.0.8
medmnist==3.0.1
numpy
pandas
scikit-learn
scikit-image
fvcore
tqdm
Pillow
fire
torchattacks
torch
torchvision
datasets 
grad-cam
seaborn
